+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - upgraded to `khao-malet@0.0.19`
 - added `label` prop to `khao-link`, slot is deprecated now!
 - improved focus state of `khao-link`
-- added hover/focus and active state to `khao-tab`
+- added bg-color hover/focus and active state to `khao-tab`
+- replace slot in `khao-tab` by `label`
 
 ## [1.9.1]